+++ src/client/Microsoft.Identity.Client.Broker/BrokerExtension.cs
@@ -4,6 +4,7 @@
 using System;
 using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.Internal.Broker;
+using Microsoft.Identity.Client.Platforms.Features.RuntimeBroker;
 using Microsoft.Identity.Client.PlatformsCommon.Shared;
 
 namespace Microsoft.Identity.Client.Broker
