+++ src/client/Microsoft.Identity.Client.Broker/Microsoft.Identity.Client.Broker.csproj
@@ -51,6 +51,8 @@
     <DefineConstants>$(DefineConstants);NET_CORE</DefineConstants>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="..\Microsoft.Identity.Client\Platforms\Features\RuntimeBroker\RuntimeBroker.cs" Link="RuntimeBroker.cs" />
+    <Compile Include="..\Microsoft.Identity.Client\Platforms\Features\RuntimeBroker\WamAdapters.cs" Link="WamAdapters.cs" />
     <Compile Include="..\Microsoft.Identity.Client\Properties\InternalsVisibleTo.cs" Link="Properties\InternalsVisibleTo.cs" />
   </ItemGroup>
     <ItemGroup Condition=" '$(TargetFramework)' == '$(TargetFrameworkNetDesktop)'">
